\name{def.grav}
\alias{def.grav}

\title{
Define a gravity system
}

\description{
This function is the basis to enter the relevant properties of a (sub-)catchment drained with open channel flow (gravity system). It can be used as a stand alone function to 
model wastewater and substance flows at the effluent of a single catchment (or premise/building) or to be linked with other sub-catchments, particularly if they are connected
with pumps (see \code{\link{def.pump}}). A "pulse" can be a toilet flush containing the substance of interest or any other wastewater pulse generated by a household appliance 
(e.g. dishwasher) containing another substance of interest.
}

\usage{
def.grav(population, frac.consumers=NULL, exp.n.pulses=NULL,
    WE.factor=1, pulse.mass, diurnal.S.WD=1, diurnal.S.WE=diurnal.S.WD,
    Q.c.day=250, diurnal.Q=1, Q.base=0, 
    flow.distance=2000, v.flow=1, Disp=0.16, pulse.dur.min=5, pulse.dur.max=95)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{population}{
    population size in the (sub-)chatchment
  }
  \item{frac.consumers}{
    fraction of population consuming substance of interest. See Details.
  }
  \item{exp.n.pulses}{
    expected absolute number of wastewater pulses per day containing the substance of interest. See Details.
  }
  \item{WE.factor}{
    factor for higher consumption on weekends (number of consumers, not amount of substance)
  }
  \item{pulse.mass}{
    a single number defining the average mass of the substance of interest discharged into
	the sewer system, or a vector from which the mass per pulse is sampled [mg]. See Details.
  }
  \item{diurnal.S.WD}{
    vector defining the diurnal distribution of the substance of interest (e.g. drug excretion) over the day (weekdays) [-]. See Details.
  }
  \item{diurnal.S.WE}{
    vector defining the diurnal distribution of the substance of interest (e.g. drug excretion) over the day (weekend) [-]. See Details.
  }
  \item{Q.c.day}{
    daily per capita wastewater production subject to diurnal variation [l/c/d]. See Details.
  }  
  \item{diurnal.Q}{
    vector defining the diurnal wastewater variation of \code{Q.c.day} [-]. See Details.
  }
   \item{Q.base}{
    constant base flow (e.g. infiltration of extraneous water) [l/s]. See Details.
  }
  \item{flow.distance}{
    a single number defining the average flow distance from the point of discharge (e.g. household
	to sewage treatment plant), or a vector containing the distribution of flow distances from which 
	will be sampled [m]. See Details.
  }
  \item{v.flow}{
    flow velocity [m/s]. See Details.
  }
  \item{Disp}{
    dispersion coefficient See Details.
  }
  \item{pulse.dur.min}{
    minimum duration of a substance pulse at the source [s]. See Details.
  }
  \item{pulse.dur.max}{
    maximum duration of a substance pulse at the source [s]. See Details.
  }
}

\details{
  Either the argument \code{frac.consumers} or \code{exp.n.pulses} must
  be specified. If \code{frac.consumers} is given, \code{exp.n.pulses}
  is calculated as the product of \code{population} and \code{frac.consumers}.

  If the distribution of mass per pulse is known, it can be entered as a vector,
  if it is not known, a single value can be entered. In the latter case, a distribution
  of masses is calculated with mean=value and sd=0.2*value (n=1000). 
  
  Diurnal variations can be defined for both, wastewater flows and occurrence of substance.
  As an approximation for wastewater flow variations, the flow at the sampling station (typically
  sewage treatment plant) can be taken shifted back by flow.distance/v.flow. A vector with 12 
  two-hourly values is typically sufficient (the sum of the vector will be normalized to 1).
  Usually, the temporal occurrence of the substance is not so well known. If pharmacokinetics are slow
  a single value of 1 (no diurnal variation expected since the substance will be excreted in similar
  amounts with each voiding) might be adequate. Alternatively, a similar diurnal pattern as for 
  wastewater flows might be appropriate reflecting the activity of the majority of people. If not 
  specified the default value for the both diurnal variations is 1 (which implies no diurnal variation).
  
  For daily wastewater volumes two values can be entered: first, a per capita value (~drinking water 
  consumption subject to diurnal variation as described above) and a constant base flow that is independent
  from water consumption (e.g. infiltration from extraneous water when sewers are below the groundwater 
  table).
  
  If the distribution of flow distances from households to the point of interest (e.g. sampling point at the
  influent of a sewage treatment plant) is known (e.g. from GIS data or a detailed hydraulic model) it can be
  entered as a vector. If it is not known, a single value can be entered (e.g. measured from Google maps) 
  and a distribution of distances is calculated with mean=value and sd=0.25*value (n=1000). 
  
  The average flow velocity can vary depending on the design of the sewers and the prevalent topography.
  To avoid the accumulation of sediments a minimum velocity is often tried to be achieved by design. 
  The default value for calculations is 1m/s (0.2-2m/s ***verify range + ref***). Slower velocities will lead 
  to longer travel times and higher dispersion and hence more attenuated short-term fluctuations of a 
  substance.
  
  While dispersion coefficients can vary (0.03-1m^2/s ***verify range + ref***) within and between sewer systems, their
  effect on the result (short-term variations and sampling) is not as sensitive as velocity, flow distance and 
  initial pulse duration (see next paragraph). A default value of 0.16 m^2/s is a value that was derived from 
  investigations summarized in Rieckermann et al. (2005).  

  
  

  For an example see the documentation of \code{\link{SPG-package}}.
}

\value{ A function to generate substance and flow pattern of a gravity
 system. It takes two arguments: \code{sim.dur}, the simulation
 duration [days] and \code{temp.res}, the temporal resolution of the
 simulation [minutes].
}

 \references{
  Ort C., Schaffner C., Giger W. and Gujer W. (2005) Modeling Stochastic Load Variations in Sewer 
  Systems. Water Science & Technology 52 (5), 113-122.
 
  Rieckermann J., Neumann M.B., Ort C., Huisman J.L. and Gujer W. (2005) Dispersion Coefficients of 
  Sewers from Tracer Experiments.  Water Science & Technology 52 (5), 123-133.
}

\author{
  Andreas Scheidegger \email{andreas.scheidegger@eawag.ch} (for R-specific coding questions)
  
  Christoph Ort \email{christoph.ort@eawag.ch} (for input data and application-related questions)
}

\seealso{
  \code{\link{SPG-package}}, \code{\link{def.pump}}, \code{\link{take.samples}}, \code{\link{plot.flow}}, \code{\link{plot.samples}}
}

% \examples{
% ## example here


% }

